requirements.txt
appnope==0.1.0
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Cors==3.0.3
flask-jsontools==0.1.1.post0
Flask-SQLAlchemy==2.3.2
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
MarkupSafe==1.0
parso==0.1.0
pexpect==4.2.1
pg8000==1.11.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.13
testing.common.database==2.0.3
testing.postgresql==1.3.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
